refactor: move authentication helpers out of ui-components
