+++ shared-helpers/src/authContext.ts
@@ -21,6 +21,7 @@ import {
   EnumRequestMfaCodeMfaType,
   EnumLoginMfaType,
 } from "@bloom-housing/backend-core/types"
+import { NavigationContext } from "@bloom-housing/ui-components"
 import {
   createContext,
   createElement,
@@ -33,10 +34,9 @@ import {
 } from "react"
 import qs from "qs"
 import axiosStatic from "axios"
-import { ConfigContext } from "../config/ConfigContext"
+import { ConfigContext } from "./ConfigContext"
 import { createAction, createReducer } from "typesafe-actions"
 import { clearToken, getToken, getTokenTtl, setToken } from "./token"
-import { NavigationContext } from "../config/NavigationContext"
 
 type ContextProps = {
   amiChartsService: AmiChartsService
