+++ shared-helpers/src/useRequireLogginInUser.ts
@@ -1,6 +1,6 @@
 import { useContext } from "react"
-import { AuthContext } from "./AuthContext"
-import { NavigationContext } from "../config/NavigationContext"
+import { AuthContext } from "./authContext"
+import { NavigationContext } from "@bloom-housing/ui-components"
 
 /**
  * Require a logged in user. Waits on initial load, then initiates a redirect to `redirectPath` if user is not
