+++ ui-components/src/config/index.ts
@@ -1,2 +1,1 @@
-export { ConfigProvider, ConfigContext } from "./ConfigContext"
 export * from "./NavigationContext"
